+++ packages/app/cypress/e2e/runner/sessions.ui.cy.ts
@@ -39,7 +39,8 @@ describe('runner/cypress sessions.ui.spec', {
 
   it('creates new session', () => {
     loadSpec({
-      filePath: 'sessions/new_session.cy.js',
+      projectName: 'runner-session-e2e-specs',
+      filePath: 'new_session.cy.js',
     })
 
     validateSessionsInstrumentPanel(['blank_session'])
@@ -62,7 +63,8 @@ describe('runner/cypress sessions.ui.spec', {
 
   it('creates new session with validation', () => {
     loadSpec({
-      filePath: 'sessions/new_session_with_validation.cy.js',
+      projectName: 'runner-session-e2e-specs',
+      filePath: 'new_session_with_validation.cy.js',
     })
 
     validateSessionsInstrumentPanel(['blank_session'])
@@ -95,7 +97,8 @@ describe('runner/cypress sessions.ui.spec', {
 
   it('creates new session and fails validation', () => {
     loadSpec({
-      filePath: 'sessions/new_session_and_fails_validation.cy.js',
+      projectName: 'runner-session-e2e-specs',
+      filePath: 'new_session_and_fails_validation.cy.js',
     })
 
     validateSessionsInstrumentPanel(['blank_session'])
@@ -127,7 +130,8 @@ describe('runner/cypress sessions.ui.spec', {
 
   it('restores saved session', () => {
     loadSpec({
-      filePath: 'sessions/restores_saved_session.cy.js',
+      projectName: 'runner-session-e2e-specs',
+      filePath: 'restores_saved_session.cy.js',
     })
 
     cy.get('.test').each(($el) => cy.wrap($el).click())
@@ -180,7 +184,8 @@ describe('runner/cypress sessions.ui.spec', {
 
   it('recreates session', () => {
     loadSpec({
-      filePath: 'sessions/recreates_session.cy.js',
+      projectName: 'runner-session-e2e-specs',
+      filePath: 'recreates_session.cy.js',
     })
 
     cy.get('.test').each(($el) => cy.wrap($el).click())
@@ -241,7 +246,8 @@ describe('runner/cypress sessions.ui.spec', {
 
   it('recreates session and fails validation', () => {
     loadSpec({
-      filePath: 'sessions/recreates_session_and_fails_validation.cy.js',
+      projectName: 'runner-session-e2e-specs',
+      filePath: 'recreates_session_and_fails_validation.cy.js',
     })
 
     cy.get('.test').each(($el) => cy.wrap($el).click())
@@ -298,7 +304,8 @@ describe('runner/cypress sessions.ui.spec', {
 
   it('multiple sessions in a test', () => {
     loadSpec({
-      filePath: 'sessions/multiple_sessions.cy.js',
+      projectName: 'runner-session-e2e-specs',
+      filePath: 'multiple_sessions.cy.js',
     })
 
     validateSessionsInstrumentPanel(['user1', 'user2'])