+++ packages/driver/cypress/e2e/commands/waiting.cy.js
@@ -68,7 +68,7 @@ describe('src/cy/commands/waiting', () => {
 
           return null
         })
-        .wait('@fetch').then((xhr) => {
+        .wait('@fetch.response').then((xhr) => {
           expect(xhr.responseBody).to.deep.eq(response)
         })
       })
@@ -302,7 +302,7 @@ describe('src/cy/commands/waiting', () => {
           .wait('getAny').then(() => {})
         })
 
-        it('throws when 2nd alias doesnt match any registered alias', (done) => {
+        it('throws when 2nd alias doesn\'t match any registered alias', (done) => {
           cy.on('fail', (err) => {
             expect(err.message).to.eq('`cy.wait()` could not find a registered alias for: `@bar`.\nAvailable aliases are: `foo`.')
 
@@ -339,7 +339,7 @@ describe('src/cy/commands/waiting', () => {
           .wait(['@foo', 'bar'])
         })
 
-        it('throws when 2nd alias isnt a route alias', (done) => {
+        it('throws when 2nd alias isn\'t a route alias', (done) => {
           cy.on('fail', (err) => {
             expect(err.message).to.include('`cy.wait()` only accepts aliases for routes.\nThe alias: `bar` did not match a route.')
             expect(err.docsUrl).to.eq('https://on.cypress.io/wait')
@@ -448,7 +448,7 @@ describe('src/cy/commands/waiting', () => {
           .wait(['@foo', 'bar'])
         })
 
-        it('does not throw again when 2nd alias doesnt reference a route', {
+        it('does not throw again when 2nd alias doesn\'t reference a route', {
           requestTimeout: 100,
         }, (done) => {
           Promise.onPossiblyUnhandledRejection(done)
@@ -1132,7 +1132,7 @@ describe('src/cy/commands/waiting', () => {
             expect(this.lastLog.invoke('consoleProps')).to.deep.eq({
               Command: 'wait',
               'Waited For': 'getFoo, getBar',
-              Yielded: [xhrs[0], xhrs[1]], // explictly create the array here
+              Yielded: [xhrs[0], xhrs[1]], // explicitly create the array here
             })
           })
         })