+++ packages/launchpad/cypress/e2e/migration.cy.ts
@@ -1,4 +1,5 @@
 import type { ProjectFixtureDir } from '@tooling/system-tests'
+import type { SinonStub } from 'sinon'
 import { getPathForPlatform } from './support/getPathForPlatform'
 
 // @ts-ignore
@@ -30,6 +31,7 @@ function scaffoldAndVisitLaunchpad (project: ProjectFixtureDir, argv?: string[])
 
 function startMigrationFor (project: ProjectFixtureDir, argv?: string[]) {
   scaffoldAndVisitLaunchpad(project, argv)
+  cy.contains('button', cy.i18n.migration.landingPage.actionContinue).click()
   cy.waitForWizard()
 }
 
@@ -40,7 +42,7 @@ function skipCTMigration () {
 function migrateAndVerifyConfig (migratedConfigFile: string = 'cypress.config.js') {
   cy.contains('Migrate the configuration for me').click()
 
-  cy.withCtx(async (ctx, o) => {
+  cy.withRetryableCtx(async (ctx, o) => {
     const configStats = await ctx.file.checkIfFileExists(o.migratedConfigFile)
 
     expect(configStats).to.not.be.null.and.not.be.undefined
@@ -90,6 +92,9 @@ describe('global mode', () => {
     })
 
     cy.contains('migration-e2e-export-default').click()
+
+    cy.contains('button', cy.i18n.migration.landingPage.actionContinue).click()
+
     // rename integration->e2e
     cy.get(renameAutoStep).should('exist')
     cy.get(renameManualStep).should('not.exist')
@@ -128,6 +133,7 @@ describe('Opening unmigrated project', () => {
     cy.scaffoldProject('migration')
     cy.openProject('migration', ['--e2e'])
     cy.visitLaunchpad()
+    cy.contains('button', cy.i18n.migration.landingPage.actionContinue).click()
     cy.get('h1').should('contain', 'Migration')
   })
 
@@ -135,7 +141,62 @@ describe('Opening unmigrated project', () => {
     cy.scaffoldProject('migration-component-testing')
     cy.openProject('migration-component-testing', ['--component'])
     cy.visitLaunchpad()
+    cy.contains('button', cy.i18n.migration.landingPage.actionContinue).click()
     cy.get('h1').should('contain', 'Migration')
+  })
+
+  it('migration landing page appears with a video', () => {
+    cy.intercept(/vimeo.com/).as('iframeDocRequest')
+    cy.intercept(/vimeocdn/).as('vimeoCdnRequest')
+    cy.scaffoldProject('migration')
+    cy.openProject('migration')
+    cy.visitLaunchpad()
+
+    cy.contains(cy.i18n.migration.landingPage.title).should('be.visible')
+    cy.contains(cy.i18n.migration.landingPage.description).should('be.visible')
+    cy.contains('button', cy.i18n.migration.landingPage.actionContinue).should('be.visible')
+    cy.contains('a', cy.i18n.migration.landingPage.linkReleaseNotes)
+    .should('be.visible')
+    .and('have.attr', 'href', 'https://on.cypress.io/changelog')
+
+    // Vimeo's implementation may change, this is just a high level check that
+    // the expected iframe code is being returned and that there is vimeo-related network traffic
+    cy.get('[data-cy="video-container"] iframe[src*=vimeo]').should('be.visible')
+    cy.wait('@iframeDocRequest')
+
+    // For an unknown reason, recaptcha blocks us from loading Vimeo on CircleCI Windows only
+    // So only wait on the Vimeo CDN request on Linux/Darwin.
+    if (Cypress.platform !== 'win32') cy.wait('@vimeoCdnRequest')
+
+    cy.percySnapshot()
+  })
+
+  it('landing page does not appear if there is no video embed code', () => {
+    cy.scaffoldProject('migration')
+    cy.openProject('migration')
+    cy.withCtx((ctx, o) => {
+      o.sinon.stub(ctx.migration, 'getVideoEmbedHtml').callsFake(async () => {
+        return null
+      })
+    })
+
+    cy.visitLaunchpad()
+    cy.contains(cy.i18n.welcomePage.title).should('be.visible')
+    cy.contains(cy.i18n.migration.landingPage.title).should('not.exist')
+  })
+
+  it('should only hit the video on link once & cache it', () => {
+    cy.scaffoldProject('migration')
+    cy.openProject('migration')
+
+    cy.visitLaunchpad()
+    cy.contains(cy.i18n.migration.landingPage.title).should('be.visible')
+
+    cy.visitLaunchpad()
+    cy.contains(cy.i18n.migration.landingPage.title).should('be.visible')
+    cy.withCtx((ctx, o) => {
+      expect((ctx.util.fetch as SinonStub).args.filter((a) => String(a[0]).includes('v10-video-embed')).length).to.eq(1)
+    })
   })
 })
 
@@ -176,8 +237,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = [
         'cypress/component/button.cy.js',
         'cypress/component/input.cy.tsx',
@@ -219,9 +279,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
       runAutoRename()
 
-      cy.wait(100)
-
-      cy.withCtx(async (ctx) => {
+      cy.withRetryableCtx(async (ctx) => {
         const specs = [
           'cypress/e2e/foo.cy.ts',
           'cypress/component/button.cy.js',
@@ -239,7 +297,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
       migrateAndVerifyConfig()
       finishMigrationAndContinue()
 
-      cy.withCtx(async (ctx) => {
+      cy.withRetryableCtx(async (ctx) => {
         const integrationFolderStats = await ctx.file.checkIfFileExists(ctx.path.join('cypress', 'integration'))
 
         expect(integrationFolderStats).to.be.null
@@ -294,9 +352,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
       cy.findByText('Rename the folder for me').click()
 
-      cy.wait(100)
-
-      cy.withCtx(async (ctx) => {
+      cy.withRetryableCtx(async (ctx) => {
         const specs = [
           'cypress/e2e/foo.spec.js',
         ]
@@ -336,9 +392,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = [
         'cypress/e2e/foo.cy.ts',
         'cypress/component/button.cy.js',
@@ -387,9 +441,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = [
         'cypress/e2e/foo.cy.ts',
         'cypress/component/button.cy.js',
@@ -407,7 +459,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
     migrateAndVerifyConfig()
     finishMigrationAndContinue()
 
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const integrationFolderStats = await ctx.file.checkIfFileExists(ctx.path.join('cypress', 'integration'))
 
       expect(integrationFolderStats).to.be.null
@@ -436,9 +488,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['src/basic.cy.js']
 
       for (const spec of specs) {
@@ -500,9 +550,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/basic.test.js']
 
       for (const spec of specs) {
@@ -547,9 +595,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/foo.cy.js']
 
       for (const spec of specs) {
@@ -594,9 +640,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/foo.spec.js']
 
       for (const spec of specs) {
@@ -641,9 +685,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/foo.spec.js']
 
       for (const spec of specs) {
@@ -687,9 +729,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/foo.cy.js']
 
       for (const spec of specs) {
@@ -725,9 +765,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/foo.cy.coffee']
 
       for (const spec of specs) {
@@ -763,9 +801,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/foo.cy.cjsx']
 
       for (const spec of specs) {
@@ -962,8 +998,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-    cy.withCtx((ctx) => {
+    cy.withRetryableCtx((ctx) => {
       ['cypress/e2e/foo.cy.js'].forEach(async (spec) => {
         const stats = await ctx.file.checkIfFileExists(ctx.path.join(spec))
 
@@ -980,7 +1015,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
     startMigrationFor('migration-specs-already-migrated')
     cy.get(renameAutoStep).should('not.exist')
 
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = [
         'cypress/tests/foo.cy.js',
       ]
@@ -1022,9 +1057,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = [
         'cypress/e2e/app-spec2.cy.js',
         'cypress/e2e/app_spec2.cy.js',
@@ -1074,9 +1107,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
       runAutoRename()
 
-      cy.wait(100)
-
-      cy.withCtx(async (ctx) => {
+      cy.withRetryableCtx(async (ctx) => {
         const specs = [
           'cypress/custom-integration/foo.cy.ts',
           'cypress/custom-component/button.cy.js',
@@ -1124,7 +1155,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
       cy.findByText('Skip renaming specs').click()
 
-      cy.withCtx(async (ctx) => {
+      cy.withRetryableCtx(async (ctx) => {
         const specs = [
           'cypress/custom-integration/foo.spec.ts',
           'cypress/custom-component/button.spec.js',
@@ -1164,6 +1195,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
     }, { path: getPathForPlatform('cypress/plugins/index.js') })
 
     cy.findByRole('button', { name: 'Try again' }).click()
+    cy.contains('button', cy.i18n.migration.landingPage.actionContinue).click()
 
     cy.waitForWizard()
   })
@@ -1198,9 +1230,7 @@ describe('Full migration flow for each project', { retries: { openMode: 0, runMo
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const files = ['cypress/e2e/foo.cy.js', 'cypress/e2e/example.json']
 
       for (const file of files) {
@@ -1285,7 +1315,7 @@ describe('Migration', { viewportWidth: 1200, retries: { openMode: 0, runMode: 2 
     cy.findByText('Rename the support file for me').click()
     cy.findByText('Migrate the configuration for me').click()
 
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const configStats = await ctx.file.checkIfFileExists('cypress.config.js')
 
       expect(configStats).to.not.be.null.and.not.be.undefined
@@ -1372,9 +1402,7 @@ describe('Migrate custom config files', () => {
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/foo.cy.js']
 
       for (const spec of specs) {
@@ -1424,9 +1452,7 @@ describe('Migrate custom config files', () => {
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/foo.cy.js']
 
       for (const spec of specs) {
@@ -1476,9 +1502,7 @@ describe('Migrate custom config files', () => {
 
     runAutoRename()
 
-    cy.wait(100)
-
-    cy.withCtx(async (ctx) => {
+    cy.withRetryableCtx(async (ctx) => {
       const specs = ['cypress/e2e/foo.cy.js']
 
       for (const spec of specs) {
@@ -1506,6 +1530,7 @@ describe('Migrate custom config files', () => {
 
   it('shows error for migration-custom-config-file-with-existing-v10-config-file', () => {
     scaffoldAndVisitLaunchpad('migration-custom-config-file-with-existing-v10-config-file', ['--config-file', 'customConfig.json'])
+    cy.contains('button', cy.i18n.migration.landingPage.actionContinue).click()
 
     cy.contains('There is both a customConfig.config.js and a customConfig.json file at the location below:')
     cy.contains('Cypress no longer supports customConfig.json, please remove it from your project.')