+++ system-tests/projects/runner-session-e2e-specs/cypress/e2e/blank_session.cy.js
@@ -1,0 +1,4 @@
+it('t1', () => {
+  cy.session('blank_session', () => {})
+  assert(true)
+})